refactor(chat): extract scroll hook into message-anchor module
Move scroll logic from zero-chat-thread-page.tsx into a reusable
hook in views/components/message-anchor.ts.
